test: consolidate quantity-field-validation tests to correct location
Move Task 14.2 tests to main test file and remove duplicate.
